refactor(sections): extract synopsis handling into usage.rs
Pure motion: recognizing a usage line or an unlabelled stanza head, the
stanza-head flag recovery, and mining the synopsis for positionals and for
flags no option table documents.
